LPS-76945 portal-configuration-settings - moves all classes to a more descriptive package path
